requirements/test.txt
-r base.txt

flake8==3.8.3
pytest==3.6.1
pytest-cov==2.5.1
pytest-flask==0.10.0
pytest-runner==4.2
Faker==0.8.15